Add files via upload
This script is designed to change the start menu layout on a windows 10 machine from a file you've created.  The script is run with our rmm agent which is where it gets those arguments.  It creates a local directory to put the file in locally and runs that local file to change the start menu.  Change things around to suit your needs if you want to use this script.